To Finalize V 0.1.0 a.1

+++ backend/feed-service/src/main/java/com/pms/feed_service/model/FeedBlock.java
@@ -15,7 +15,7 @@ public class FeedBlock {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Integer feed_block_id;
     private Integer placement_id;
-    private Integer block_quantity;
+    private double block_quantity;
     @ManyToOne
     @JoinColumn(name = "feed_id")
     private FeedStorage feed_storage;
